feat: implement layered execution for provisioning commands with matrix strategy
