+++ .github/workflows/setup-go.yml
@@ -6,6 +6,12 @@ on:
 jobs:
   setup-go:
     runs-on: macos-15
+    strategy:
+      fail-fast: false
+      matrix:
+        tag:
+          - go-platform
+          - go-tools
     steps:
       - name: Checkout repository
         uses: actions/checkout@v5
@@ -16,5 +22,5 @@ jobs:
       - name: Setup Ansible environment
         uses: ./.github/actions/setup-ansible
 
-      - name: Setup go
-        run: mise exec -- just run make go
+      - name: Run ${{ matrix.tag }} setup
+        run: mise exec -- just run make ${{ matrix.tag }}
