+++ .github/workflows/setup-rust.yml
@@ -6,6 +6,12 @@ on:
 jobs:
   setup-rust:
     runs-on: macos-15
+    strategy:
+      fail-fast: false
+      matrix:
+        tag:
+          - rust-platform
+          - rust-tools
     steps:
       - name: Checkout repository
         uses: actions/checkout@v5
@@ -16,5 +22,5 @@ jobs:
       - name: Setup Ansible environment
         uses: ./.github/actions/setup-ansible
 
-      - name: Setup rust
-        run: mise exec -- just run make rust
+      - name: Run ${{ matrix.tag }} setup
+        run: mise exec -- just run make ${{ matrix.tag }}
